Added authentication and user check for delete.

+++ codalab/apps/api/views.py
@@ -16,6 +16,7 @@ class CompetitionAPIViewSet(viewsets.ModelViewSet):
     serializer_class = serializers.CompetitionSerial
     queryset = webmodels.Competition.objects.all()
 
+    @action(permission_classes=[permissions.IsAuthenticated])
     def destroy(self, request, pk):
         """
         Cleanup the destruction of a competition.
@@ -25,10 +26,16 @@ class CompetitionAPIViewSet(viewsets.ModelViewSet):
         """
         # Get the competition
         c = webmodels.Competition.objects.get(id=pk)
-        c.delete()
 
-        # for each phase, cleanup the leaderboard and submissions
-        return Response(json.dumps({'id' : pk}), content_type="application/json")
+        # Create a blank response
+        response = {}
+        if self.request.user == c.creator:
+            c.delete()
+            response['id'] = pk
+        else:
+            response['status'] = 403
+
+        return Response(json.dumps(response), content_type="application/json")
 
     @action(permission_classes=[permissions.IsAuthenticated])
     def participate(self,request,pk=None):
@@ -64,7 +71,6 @@ class CompetitionAPIViewSet(viewsets.ModelViewSet):
 
     @action(methods=['POST','PUT'], permission_classes=[permissions.IsAuthenticated])
     def participation_status(self,request,pk=None):
-        print "made it into handler"
         comp = self.get_object()
         resp = {}
         status = request.DATA['status']
